refactored groupAndStackData and groupData files to dry out code
